# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Departamento de Sistemas",
        "description": "Documentación de la API para el sistema de gestión del Departamento de Sistemas.",
        "contact": {
            "email": "soporte@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/pc-caracteristicas-tecnicas": {
            "get": {
                "tags": [
                    "PcCaracteristicasTecnicas"
                ],
                "summary": "Listar características técnicas",
                "operationId": "952ddc7986ddf10ca25607c3ad628d50",
                "responses": {
                    "200": {
                        "description": "Lista obtenida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PcCaracteristicasTecnicas"
                ],
                "summary": "Crear características técnicas",
                "operationId": "2da779f614256631e1aa68b8679790ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "equipo_id": {
                                        "type": "integer"
                                    },
                                    "procesador": {
                                        "type": "string"
                                    },
                                    "memoria_ram": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pc-caracteristicas-tecnicas/{id}": {
            "get": {
                "tags": [
                    "PcCaracteristicasTecnicas"
                ],
                "summary": "Obtener por ID",
                "operationId": "0bbe8af0969c9a26eb7211f7d4a48347",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "PcCaracteristicasTecnicas"
                ],
                "summary": "Actualizar características",
                "operationId": "50cfb964096e335853c11cd79c791465",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "procesador": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "PcCaracteristicasTecnicas"
                ],
                "summary": "Eliminar características",
                "operationId": "e9662ffb4069cdc7a9ec9814570f2d97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pc-caracteristicas-tecnicas/equipo/{equipo_id}": {
            "get": {
                "tags": [
                    "PcCaracteristicasTecnicas"
                ],
                "summary": "Obtener por ID de Equipo",
                "operationId": "3f8c8f8ab9ab3e9920912e2359dc14ca",
                "parameters": [
                    {
                        "name": "equipo_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "title": "Respuesta Estándar",
                "description": "Estructura de respuesta estándar para la API",
                "properties": {
                    "mensaje": {
                        "type": "string",
                        "example": "Operación exitosa"
                    },
                    "objeto": {
                        "type": "object",
                        "example": []
                    },
                    "status": {
                        "type": "integer",
                        "example": 200
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT en el formato: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}